Added a flag to toggle concurrency generation of segments

+++ helpers/prompt.bash
@@ -87,14 +87,20 @@ function generate_prompt() {
   local prompt_right_end=$(( ${#settings_segments_right[@]} + prompt_left_end ))
   local prompt_segments=( ${settings_segments_left[@]} ${settings_segments_right[@]} ${settings_segment_line_two[@]} )
 
-  # Concurrent evaluation of promt segments
-  tempdir=$(mktemp -d) && trap 'rm -rf "$tempdir"' EXIT;
-  for i in "${!prompt_segments[@]}"; do
-    generate_segment_value "${prompt_segments[i]}" "$exit_code" "$command_time" > "$tempdir/$i" & pids[i]=$!
-  done
-  for i in "${!pids[@]}"; do
-    wait "${pids[i]}" && prompt_segments[i]=$(<"$tempdir/$i");
-  done
+  if [[ "$settings_concurrency_enabled" == true ]]; then
+    # Concurrent evaluation of promt segments
+    tempdir=$(mktemp -d) && trap 'rm -rf "$tempdir"' EXIT;
+    for i in "${!prompt_segments[@]}"; do
+      generate_segment_value "${prompt_segments[i]}" "$exit_code" "$command_time" > "$tempdir/$i" & pids[i]=$!
+    done
+    for i in "${!pids[@]}"; do
+      wait "${pids[i]}" && prompt_segments[i]=$(<"$tempdir/$i");
+    done
+  else
+    for i in "${!prompt_segments[@]}"; do
+      prompt_segments["$i"]=$(generate_segment_value "${prompt_segments[i]}" "$exit_code" "$command_time")
+    done
+  fi
 
   # Format the segments
   local previous_segment=
